app/requirements.txt
streamlit
requests
pandas
numpy
plotly
kaleido
tenacity
pydantic
python-pptx
openpyxl
reportlab
